% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artuts.R
\name{artuts}
\alias{artuts}
\title{artuts Function}
\usage{
artuts()
}
\description{
This is a convience function that returns a list
with article level information from the nttpr.private private utilities
for return message, execution time stats, and article info
with starting and ending article numbers and their respective article dates.
}
