% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect Function}
\usage{
connect(nntp_server, nntp_port, user_name, user_password, read_only)
}
\arguments{
\item{nntp_server}{dsn name or ip address of target NNTP server}

\item{nntp_port}{connect port of NNTP server}

\item{user_name}{user login name}

\item{user_password}{user login password}

\item{read_only}{sets read only session True or False}
}
\description{
This function creates an NNTP connection object.
An NNTP connection object is and either the welcome message or an error message
is stored in nntpr.private.
}
