% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getanarticle.R
\name{getanarticle}
\alias{getanarticle}
\title{getanarticle Function}
\usage{
getanarticle(article_string, request_type)
}
\arguments{
\item{article_string}{article number to retrieve}

\item{request_type}{the nntplib method to execute (article, head, body)}
}
\description{
This function retrieves a single article header and body,
header, or body (depending on request type parameter)
from the group set by the last group() call.
The function returns 4 item list with:
\itemize{
\item request summary with req number, article number, and message id
\item article number string
\item message id
\item vector containing header, body, or header and body
}
or and error message string if applicable.
}
