% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{group Function}
\usage{
group(groupname)
}
\arguments{
\item{groupname}{usenet group name}
}
\description{
This function gets group information and sets the group for nntplib methods operating against articles.
Group information vector (response, article count, first article number, last article number, group name)
is stored in nntpr.private and is returned to caller.
}
