% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupdatevolume.R
\name{groupdatevolume}
\alias{groupdatevolume}
\title{groupdatevolume Function}
\usage{
groupdatevolume(runmode = NULL)
}
\arguments{
\item{runmode}{optional parameter, set to 'test' request a smaller sample set for testing purposes}
}
\description{
This function provides a timevis timeline plot, providing an overview of
a newsgroup's average posts per day, ranked across quartiles, for a series of
date ranges from first to last available article. A sample size is calculated, determining
the number of article dates retrieved from usenet at mostly equal article count distances.
The date ranges plotted represent the resulting sample periods, with a calculated average
number of posts per day for that sample period.
}
\details{
groupdatevolume calls a private function, samplegroup, to collect
the article post dates for a randomly generated sample set of article numbers,
returning a data frame of the sample set data carrying article number, post date, and for convenience
an article count since previous sample point.

The data is then sorted by posts per day and assigned color codes across the sample's
posts per day quartiles.  That is passed to the timevis package to create an interactive
time line plot.  The time line plot object is returned to the caller.

The default margin of error used to calculate sample size (confidence interval) is 5%, but calling
groupdatevolume with 'test' as a parameter will change that to 20%.  This returns a sample size
one quarter of the default size.
}
