% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listartheader.R
\name{listartheader}
\alias{listartheader}
\title{listartheader Function}
\usage{
listartheader(first_article_string, last_article_string, timeout_seconds,
  filter_term_string = NULL)
}
\arguments{
\item{first_article_string}{article number (string)}

\item{last_article_string}{number (string)}

\item{timeout_seconds}{number of seconds to set withTimeout to, 0 skips withTimeout}

\item{filter_term_string}{string (optional, case insensitive)}
}
\description{
This function gets a list of article headers from newsgroup set by last group call.
If an error is encountered, the error message is stored in nntpr.private, and NULL is returned
to the caller, otherwise the dates of the first article and last article
in the request are stored in nntpr.private and the article list returned to the caller.
}
