% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listgroups.R
\name{listgroups}
\alias{listgroups}
\title{listgroups Function}
\usage{
listgroups(wildcard = NULL)
}
\arguments{
\item{wildcard}{optional usenet group name search wildcard}
}
\description{
This function retrieves a list of groups (with an optional search wildcard string)
and returns a group list to the caller.
}
