% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listheaderfield.R
\name{listheaderfield}
\alias{listheaderfield}
\title{listheaderfield Function}
\usage{
listheaderfield(header_field_name, first_article_string,
  last_article_string, timeout_seconds, filter_term_string = NULL)
}
\arguments{
\item{header_field_name}{header field name (path, from, subject, newsgroups, user-agent, message-id, x-no-archive, lines, x-complaints-to, nntp-posting-date, organization, bytes, date, x-received-bytes, x-received-body-crc)}

\item{first_article_string}{number (string)}

\item{last_article_string}{number (string)}

\item{timeout_seconds}{number of seconds to set withTimeout to, 0 skips withTimeout}

\item{filter_term_string}{string (optional, case insensitive)}
}
\description{
This function gets a list of article header requested field values from newsgroup set by last group call.
If an error is encountered, the error message is stored in nntpr.private, and NULL is returned
to the caller, otherwise the dates of the first article and last article
in the request are stored in nntpr.private and the article header field values list returned to the caller.
}
