% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntpr.quit.R
\name{nntpr.quit}
\alias{nntpr.quit}
\title{nntpr.quit Function}
\usage{
nntpr.quit()
}
\description{
This function disconnects the NNTP session associated with a connection object.
The 'goodby' message or a warning that there's nothing to quit from is stored
in nntpr.private and is returned to caller.
}
