% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{nntpr}
\alias{nntpr}
\alias{nntplib}
\alias{nntpr-package}
\title{nntpr Function}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
nntplib
}
\description{
This function initializes the R reticulate environment for the rest of nntpr
}
\keyword{internal}
