% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posttext.R
\name{posttext}
\alias{posttext}
\title{posttext Function}
\usage{
posttext(group_list, post_subject, message_body_vector)
}
\arguments{
\item{group_list}{values for the Newsgroups header}

\item{post_subject}{value for the Subject header}

\item{message_body_vector}{vector containing lines of message text to post}
}
\description{
This function posts a text message
with whatever header information you have stored
in nntpr.private, and the list of groups, the subject,
and text body passed in function parameters
}
