% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicutility.R
\name{publicutility}
\alias{publicutility}
\title{publicutility Function}
\usage{
publicutility(utility_to_call)
}
\arguments{
\item{utility_to_call}{the function you want to invoke}
}
\description{
This function provides a public method to
 interface with private utility functions.
Returns whatever the private function returns.
}
\details{
Valid utility to call:
\itemize{
\item help - return a message with valid utility to call values
\item retmessage - return nntpr.private$gretmessage
\item artinfo - return 2 item list with first-last article info from last listarticles call
\item groupinfo - return vector with Group information vector (response, count, first, last, name)
\item exectime - return string with last stored system.time information collected on nntplib xhdr and xover calls
}
}
