% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_post_names.R
\name{set_post_names}
\alias{set_post_names}
\title{set_post_names Function}
\usage{
set_post_names(from_name, email_name, email_domain)
}
\arguments{
\item{from_name}{sender name displayed in the NNTP From header}

\item{email_name}{sender email box name (before the @) for the NNTP From header}

\item{email_domain}{sender email address domain (after the @) for the NNTP From and Message-ID headers}
}
\description{
This function stores the name, email box name, and email domain
in nntpr.private for use by posttext in constructing the
NNTP message headers.
Calling posttext before using this to set the names
will return an error condition.
}
